Upgraded EventAggregation_Desktop solution to Prism 6.1.
Not all tests are passing.

If line 52 in  ActivityPresenter.cs is set to      subscriptionToken = fundAddedEvent.Subscribe(FundAddedEventHandler, ThreadOption.UIThread, false, FundOrderFilter);

ActivityPresenter Fixture tests fail with

"To use the UIThread option for subscribing, the EventAggregator must be constructed on the UI thread."

If line 52 in  ActivityPresenter.cs is set to      subscriptionToken = fundAddedEvent.Subscribe(FundAddedEventHandler, ThreadOption.PublisherThread, false, FundOrderFilter);

ActivityPresenter Fixture tests pass but other tests fail.

This could be a by product of MSTest? Not sure really.

Checking the code in anyway since it is almost readyl